perf: modernize component example documentation #110 #111 #112 example code updated to reflect latest Auro element APIs example documentation layout improved fix broken unformatted drawer example custom registration updated and moved to install page properties now listed alphabetically doc build system now uses the latest version of auro-cli package json configuration updated to support latest auro-cli additional packages dependencies updated to latest versions
