+++ apiExamples/close-button-appearance.html
