+++ apiExamples/close-button-appearance.js
