+++ docs/partials/description.md
@@ -1,3 +1,3 @@
-The `<auro-drawer>` component is an intrusive interactive component, not passive. The component is best used when there is a requirement the user be made aware of the content being shown, moving focus from the main content to the drawer content.
+The `<auro-drawer>` element is an intrusive interactive element, not passive. The element is best used when there is a requirement the user be made aware of the content being shown, moving focus from the main content to the drawer content.
 
-The component also supports a modal (blocking) state where the user must interact with the content of the component in order to continue. Passive dismissal of the content is not allowed. Users of this state must add a trigger for the user to move beyond this content.
+The element also supports a modal (blocking) state where the user must interact with the content of the element in order to continue. Passive dismissal of the content is not allowed. Users of this state must add a trigger for the user to move beyond this content.
