--- scripts/wca/auro-drawer.js
@@ -1,10 +1,0 @@
-import { AuroDrawer } from "../../src/auro-drawer.js";
-
-/**
- * @private
- */
-class AuroDrawerWCA extends AuroDrawer {}
-
-if (!customElements.get("auro-drawer")) {
-  customElements.define("auro-drawer", AuroDrawerWCA);
-}
